Remove final sleep call and fix broken tests
