--- src/utils/sleep.ts
@@ -1,5 +1,0 @@
-export const sleep = async (ms: number): Promise<void> => {
-  return new Promise((resolve) => {
-    setTimeout(resolve, ms);
-  });
-};
